+++ app/src/main/java/rs/ltt/android/entity/EmailEntity.java
@@ -34,7 +34,7 @@ public class EmailEntity {
 
     public String threadId;
 
-    public Integer size;
+    public Long size;
 
     public Date receivedAt;
